Create read_stdin_write_to_file.c
Actual Coding Challenge  

--------------------------------

"Can you write a program in C (not C++) that reads arbitrary data from standard input and writes it to a file, please? 

By arbitrary data, we mean that we don’t want to put any limitations or restrictions whatsoever on the data a user provides via standard input.